Refactor plots into submodule.
